+++ HISTORY.md
@@ -9,6 +9,7 @@
 ### 2021-04-12
 * blocks, objects, threads, gui: new "switch to scene _" command primitive
 * morphic, gui: support bulk-file-drop for importing scenes
+* gui: tweaked scene album colors
 
 ### 2021-04-08
 * gui: scroll selected scene icon into view